test(roborazzi): promote settings dialogs/downloads to committed goldens [skip changelog]
