+++ docs/screenshots/README.md
@@ -1,45 +1,49 @@
 # Screenshot baselines
 
-This document describes the current screenshot and visual-regression reference path. Screenshot comparison is not a required CI gate at this time.
+This document describes the committed Roborazzi goldens and the visual-regression reference path. `:feature:home:verifyRoborazziDebug` runs in merge CI.
 
 ## Current coverage
 
-- No PNG goldens are committed under `screenshots/baselines/`.
-- `:feature:home` contains `AddRssFeedDialogRoborazziSpikeTest`, which can capture a local Roborazzi image under `feature/home/build/outputs/roborazzi/`.
-- CI does not run image diff comparisons.
-- Home instrumented tests verify dialog composition through stable Compose test tags.
+`:feature:home` hosts JVM Roborazzi golden tests in `SettingsGoldenRoborazziTest`, and their committed baselines live under `screenshots/baselines/`:
+
+```text
+screenshots/baselines/add_rss_feed_dialog.png
+screenshots/baselines/reset_analytics_dialog.png
+screenshots/baselines/downloads_settings_page.png
+```
+
+Home instrumented tests additionally verify dialog composition through stable Compose test tags.
 
 ## Layout
 
 ```text
-screenshots/baselines/     # committed PNG goldens when visual baselines are adopted
+screenshots/baselines/     # committed PNG goldens (Roborazzi output dir)
 docs/screenshots/          # this reference document
 ```
 
-Suggested names for future goldens:
-
-```text
-screenshots/baselines/add_rss_feed_dialog.png
-screenshots/baselines/home_settings_entry.png
-```
+The output directory and file naming are configured in `feature/home/build.gradle.kts` (`roborazzi { outputDir = screenshots/baselines }`) and `gradle.properties` (`roborazzi.record.filePathStrategy=relativePathFromRoborazziContextOutputDirectory`), so the short names passed to `captureRoboImage()` resolve directly under `screenshots/baselines/`.
 
 ## Capture
 
-Run the local Roborazzi capture test from the repository root:
+Regenerate the goldens from the repository root and commit any intended changes:
 
 ```bash
-./gradlew :feature:home:testDebugUnitTest --tests '*AddRssFeedDialogRoborazziSpikeTest'
+./gradlew :feature:home:recordRoborazziDebug
 ```
 
-The generated images are local build artifacts. For device captures, use a fixed-density emulator, capture the intended screen manually, and commit PNGs under `screenshots/baselines/` only when they are approved baselines.
+Goldens are captured with a fixed `lightColorScheme()` so they stay deterministic across machines. Prefer tagged-node assertions for dialogs because Compose `AlertDialog` can produce multiple roots.
 
 ## Compare
 
-Until a comparison task is wired into CI, review image changes locally with an image diff viewer or through normal pull-request artifact review. Prefer tagged-node assertions for dialogs because Compose `AlertDialog` can produce multiple roots.
+```bash
+./gradlew :feature:home:verifyRoborazziDebug
+```
+
+Verify re-renders each screen and compares it against the committed baseline; mismatches write `*_compare.png` under `feature/home/build/outputs/roborazzi/` and fail the task.
 
 ## Gradle
 
-No screenshot comparison task is part of required CI. Roborazzi is currently configured for local Home module capture tests.
+`:feature:home:verifyRoborazziDebug` is wired into `.github/workflows/unit-tests.yml` (merge CI). Roborazzi is applied through the `roborazzi` plugin alias in `feature/home/build.gradle.kts`.
 
 ## See also
