+++ gradle.properties
@@ -5,3 +5,6 @@ android.useAndroidX=true
 android.enableJetifier=false
 kotlin.code.style=official
 org.gradle.jvmargs=-Xmx2048m -Dfile.encoding=UTF-8
+# Resolve relative captureRoboImage() paths against the Roborazzi output dir
+# (screenshots/baselines) instead of each module's working directory.
+roborazzi.record.filePathStrategy=relativePathFromRoborazziContextOutputDirectory
